Tag the RosBE-Unix 0.3.6 release

+++ reactos/base/setup/usetup/fslist.c
@@ -138,7 +138,7 @@ DrawFileSystemList(
     PLIST_ENTRY ListEntry;
     PFILE_SYSTEM_ITEM Item;
     COORD coPos;
-    DWORD Written;
+    ULONG Written;
     ULONG Index = 0;
     CHAR Buffer[70];
